requirements.txt
playwright
pytest
pytest-bdd
pytest-xdist
allure-pytest
requests
faker
PyYAML

